+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/repository/rest/irs/model/response/Parameter.java
@@ -1,4 +1,0 @@
-package org.eclipse.tractusx.traceability.assets.infrastructure.repository.rest.irs.model.response;
-
-public record Parameter(String direction) {
-}