+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/infrastructure/alert/model/AlertNotificationEntity.java
@@ -32,7 +32,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import org.eclipse.tractusx.traceability.assets.infrastructure.model.AssetAsBuiltEntity;
+import org.eclipse.tractusx.traceability.assets.infrastructure.asbuilt.model.AssetAsBuiltEntity;
+import org.eclipse.tractusx.traceability.assets.infrastructure.asplanned.model.AssetAsPlannedEntity;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationAffectedPart;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationMessage;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.model.QualityNotificationStatus;
@@ -64,6 +65,13 @@ public class AlertNotificationEntity extends QualityNotificationMessageBaseEntit
     )
     private List<AssetAsBuiltEntity> assets;
 
+    @ManyToMany(cascade = CascadeType.ALL)
+    @JoinTable(
+            name = "asset_as_planned_alert_notifications",
+            joinColumns = @JoinColumn(name = "alert_notification_id"),
+            inverseJoinColumns = @JoinColumn(name = "asset_id")
+    )
+    private List<AssetAsPlannedEntity> assetsAsPlanned;
 
     public static QualityNotificationMessage toDomain(AlertNotificationEntity alertNotificationEntity) {
         return QualityNotificationMessage.builder()