+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Spring profiles for integration int-a and int-b environments
 - Quality Alert feature
 - Updated testdata to reflect asPlanned assets
+- Added asPlanned lifecycle to the asset domain
 - Added Semantic Data Model 'PartAsPlanned'
 
 ### Changed